+++ src/main/scala/org/bykn/buildcart/Hash.scala
@@ -1,0 +1,7 @@
+package org.bykn.buildcart
+
+/**
+ * An opaque value which we use to check equality
+ * on A types
+ */
+trait Hash[A]
