+++ src/main/scala/org/bykn/buildcart/Hashable.scala
@@ -1,0 +1,13 @@
+package org.bykn.buildcart
+
+/**
+ * A typeclass for hashing values.
+ */
+trait Hashable[A] {
+  def hash(a: A): Hash[A]
+}
+
+object Hashable {
+  def hash[A](a: A)(implicit h: Hashable[A]): Hash[A] =
+    h.hash(a)
+}
